Ajout du certificat intermédiaire dans les tests

+++ lib/RgsCertificate.class.php
@@ -19,21 +19,29 @@ class RgsCertificate {
 		return $this->last_message;
 	}
 
-	public function isRgsCertificate($x509_pem_certificate,$clientCertChain){
+	public function isRgsCertificate($x509_pem_certificate,$clientCertChain =null){
         $indice = mt_rand(0, mt_getrandmax());
         $tmp_cert = "/tmp/s2low-lib-rgscertificate-". $indice .".pem";
 		file_put_contents($tmp_cert,$x509_pem_certificate);
 
-		$tmp_chain = "/tmp/s2low-lib-certchain-". $indice .".pem";
-        file_put_contents($tmp_chain,$clientCertChain);
+		if(!is_null($clientCertChain)) {
+            $tmp_chain = "/tmp/s2low-lib-certchain-" . $indice . ".pem";
+            file_put_contents($tmp_chain, $clientCertChain);
+
+            $command = "{$this->openssl_path} verify -verbose -CApath {$this->validca_path} -untrusted {$tmp_chain} {$tmp_cert} 2>&1";
+        }
+		else {
+            $command = "{$this->openssl_path} verify -verbose -CApath {$this->validca_path} {$tmp_cert} 2>&1";
+        }
 
 
 		//Il semble qu'il n'y a pas de fonction php openssl_* qui permettent la vérification d'un certificat
-		$command = "{$this->openssl_path} verify -verbose -CApath {$this->validca_path} -untrusted {$tmp_chain} {$tmp_cert} 2>&1";
 
 		exec($command,$output,$return_var);
 		unlink($tmp_cert);
-		unlink($tmp_chain);
+		if(isset($tmp_chain)) {
+            unlink($tmp_chain);
+        }
 
 
 		$output = implode("\n",$output);
